Add file generated by python webserver 2 gitignore
The project suggests using an inbuilt feature of python (which is typically pre-installed on all OSX and LINUX builds) to run a web server which can then be used to view the static pages.

However when you do this it does generate a file in the project repo typically named `root@your_ip_address`. This change simply updates the `.gitignore` file to ensure it is ignored.