Update WEX static pages to embed all assets
In recent testing of the WEX enviroment it was found that when Nginx returns its static "service unavailable page" it has no styling or images. Turns out it doesn't seem to like referencing and serving up linked assets.

So to get around this problem this change updates the WEX static pages so

- all CSS is included within the page
- all images are embedded as base64 values

With these changes the html is a complete standalone file and as such should actually match the service.

One thing to note; we chose not to embed the fonts because there are a number referenced in a standard page, and when encoded to base64 they are very large and cause the html page to become massive.