Update Water Abstraction static pages
First thing we have done is rename the files to use the prefix which matches our AWS environments! The new dev team thought we didn't have any maintenance pages because no one recognised `mywail`.

We certainly needed to update the back office ones. The previous team must have used Flood Risk Activity Exemptions as its template because the look and feel was completly different to what you actuall see in the Water Abstraction back office.

Also, it used the old style of referencing assets in the `/assets` folder. Later static pages now have everything embedded in the page so no external assets are needed. This works better for web-ops and services when it comes to maintenance pages.

So, after renaming the files we've updated them based on the WEX pages. With these changes the Water Abstraction Licence service pages should be pretty up to date.